cxmanage: Change fwupdate --slot argument, and version behavior
The new options for --slot are: OLDEST, NEWEST, FIRST, SECOND, BOTH

FIRST, SECOND, and BOTH are equivalent to the ones from before, but
I think their purpose is clearer this way. OLDEST and NEWEST pick
slots based on partition versions. OLDEST is the default, since it's
the least likely to overwrite an active partition.

Additionally, when updating an image with no version specified,
cxmanage will raise the version number so that the new partition is
loaded by default, regardless of the slot's location.