fwupdate: Don't reset the MC
Multi-node fabric may not handle the "mc reset" command well, since
the nodes have to come up together. We'll want to add this back in
once we know we can do it safely.

For now, either reset the power on the chassis manually, or use the
"cxmanage mcreset" command if you trust it enough.