cxmanage: Add "inactive" and "third" slot options
Inactive is the new default, and will prioritize updating to slots
that are known to not be in use. For slots whose "in use" status is
unknown, it will update to the oldest slot.

The third option is added in case anybody really feels like
updating a running CDB partition. Good luck with that.