cxmanage: Remove "fwupdate spif" cmd, add "spifupdate" to plugin
We don't want the unreliable spi_flash updates to be an option in
the external release, so make it internal only.

I still don't recommend using it. You'll be much better off if you
can update with the .tar.gz packages. But I'm leaving it as an
option if you really have no alternatives.

You can run it like this:

cxmanage spifupdate spi_flash.bin 192.168.100.250