cxmanage: Add more informative output
The first here is to make cxmanage indicate, up front, what it is
doing -- mainly, what hosts are being operated on.

Second goal is to make the indicator more meaningful while the
command is in progress. We can easily print the number of nodes,
successes, and errors encountered so far.