cxmanage: Add option to read hosts from a hostfile
This is specified by using a "hostfile=" entry in the hostname
argument. For example, to run the info command using a file called
"example_file":

cxmanage info hostfile=example_file

The hostfile itself can just be a simple list of entries, separated
by spaces, newlines, or commas. The # character can be used for
comments.