cxmanage: Add --retry COUNT option
This option will cause cxmanage to retry the command a fixed number
of times on failed hosts, instead of prompting. This will probably
be useful for very large clusters where failures are more likely to
occur.

For example, to try the config reset command 3 times (2 retries):

cxmanage --retry 2 config reset