CXMAN-151: Add a prompt to verify the node count
This prompt only applies to the "fwupdate" and "config reset"
commands. This is so the user can make sure we got the expected
node count from ipinfo.

The -n option can be used to provide the expected number of nodes
and skip the prompt, I.E. "cxmanage -a -n 48 fwupdate ..."

The check/prompt can also be skipped entirely with the --force
option.