Moved tftp to lazy init.
When we load tests from modules, each testcase gets initialized,
therefore a fabric is created for every test case based on the manifest.

This caused a problem with LTP which generated ~1000 tests, thus ~1000
fabrics and finally ~1000 Internal tftp servers.
Move tftp to lazy init so we only create one when we need one, on
demand.