CXMAN-255: Retry config reset up to 3 times
We can't really guarantee that the CDB clear has occurred unless
the IPMI command returns properly, so, we better try it a couple
times.

This could cause a few unnecessary SPI erases in some cases, but
it's better than potentially reporting success when it hasn't
actually been reset at all.