CXMAN-277: Remove guid from node.__hash__ and node.__eq__
Since we need an IPMI command to get the guid. Having it in __hash__
can cause weird behavior -- for example, as soon as we add the Node
object to a set, we may fire off an IPMI command. Which could raise an
exception.

We can still make use of GUIDs in fabric.refresh though.