nojira: Clean up pylint warnings in cxmanage_api.tests
I moved a bunch of the in-line pylint suppression to the top of the
modules instead. I find the comments get in the way of readability,
which sort of defeats the whole purpose.

Doing module-level suppression at least gets those comments out of
the way, but isn't ideal either. We really should just leave the
warnings there.